refactor: split component files
